+++ src/Publisher.php
@@ -14,9 +14,11 @@ declare(strict_types=1);
 namespace Symfony\Component\Mercure;
 
 use Symfony\Component\HttpClient\HttpClient;
-use Symfony\Component\Mercure\Jwt\CallableTokenProvider;
-use Symfony\Contracts\HttpClient\Exception\ExceptionInterface;
+use Symfony\Component\Mercure\Internal\QueryBuilder;
+use Symfony\Component\Mercure\Jwt\TokenProviderInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
+
+trigger_deprecation('symfony/mercure', '0.5', 'Class "%s" is deprecated, use "%s" instead.', Publisher::class, Hub::class);
 
 /**
  * Publishes an update to the hub.
@@ -24,40 +26,26 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
  * @author Kévin Dunglas <dunglas@gmail.com>
  *
  * @experimental
+ *
+ * @deprecated
  */
 final class Publisher implements PublisherInterface
 {
-    private $hub;
+    private $hubUrl;
+    private $jwtProvider;
     private $httpClient;
 
     /**
-     * @param Hub                      $hub
-     * @param HttpClientInterface|null $httpClient
+     * @param (callable(Update $update):string)|TokenProviderInterface $jwtProvider
      */
-    public function __construct($hub, $httpClient = null)
+    public function __construct(string $hubUrl, $jwtProvider, HttpClientInterface $httpClient = null)
     {
-        if (!$hub instanceof Hub) {
-            trigger_deprecation('symfony/mercure', '0.5', 'Passing a hub url, and a callable to "%s::__construct()" is deprecated, pass a "%s" instance instead.', __CLASS__, Hub::class);
-
-            $hub = new Hub((string) $hub, new CallableTokenProvider($httpClient));
-            $httpClient = \func_get_args()[2] ?? null;
-        }
-
-        $this->hub = $hub;
+        $this->hubUrl = $hubUrl;
+        $this->jwtProvider = $jwtProvider;
         $this->httpClient = $httpClient ?? HttpClient::create();
     }
 
     public function __invoke(Update $update): string
-    {
-        trigger_deprecation('symfony/mercure', '0.5', 'Method "%s()" is deprecated, use "%s::publish()" instead.', __METHOD__, __CLASS__);
-
-        return $this->publish($update);
-    }
-
-    /**
-     * Sends $update to the mercure hub.
-     */
-    public function publish(Update $update): string
     {
         $postData = [
             'topic' => $update->getTopics(),
@@ -68,17 +56,17 @@ final class Publisher implements PublisherInterface
             'retry' => $update->getRetry(),
         ];
 
-        $jwt = $this->hub->getProvider()->getJwt();
+        if ($this->jwtProvider instanceof TokenProviderInterface) {
+            $jwt = $this->jwtProvider->getJwt();
+        } else {
+            $jwt = ($this->jwtProvider)($update);
+        }
         $this->validateJwt($jwt);
 
-        try {
-            return $this->httpClient->request('POST', $this->hub->getUrl(), [
-                'auth_bearer' => $jwt,
-                'body' => Internal\QueryBuilder::build($postData),
-            ])->getContent();
-        } catch (ExceptionInterface $exception) {
-            throw new Exception\RuntimeException('Failed to send an update.', 0, $exception);
-        }
+        return $this->httpClient->request('POST', $this->hubUrl, [
+            'auth_bearer' => $jwt,
+            'body' => QueryBuilder::build($postData),
+        ])->getContent();
     }
 
     /**