+++ tests/PublisherTest.php
@@ -14,20 +14,18 @@ declare(strict_types=1);
 namespace Symfony\Component\Mercure\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\HttpClient\Exception\TransportException;
 use Symfony\Component\HttpClient\MockHttpClient;
 use Symfony\Component\HttpClient\Response\MockResponse;
 use Symfony\Component\Mercure\Exception\InvalidArgumentException;
-use Symfony\Component\Mercure\Exception\RuntimeException;
-use Symfony\Component\Mercure\Hub;
 use Symfony\Component\Mercure\Jwt\StaticJwtProvider;
-use Symfony\Component\Mercure\Jwt\StaticTokenProvider;
 use Symfony\Component\Mercure\Publisher;
 use Symfony\Component\Mercure\Update;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
 /**
  * @author Kévin Dunglas <dunglas@gmail.com>
+ *
+ * @group legacy
  */
 class PublisherTest extends TestCase
 {
@@ -46,36 +44,6 @@ class PublisherTest extends TestCase
             return new MockResponse('id');
         });
 
-        $provider = new StaticTokenProvider(self::JWT);
-        $hub = new Hub(self::URL, $provider);
-        $publisher = new Publisher($hub, $httpClient);
-        $id = $publisher->publish(new Update(
-            'https://demo.mercure.rocks/demo/books/1.jsonld',
-            'Hi from Symfony!',
-            true,
-            'id',
-            null,
-            3
-        ));
-
-        $this->assertSame('id', $id);
-    }
-
-    /**
-     * @group legacy
-     */
-    public function testPublishLegacy()
-    {
-        $httpClient = new MockHttpClient(function (string $method, string $url, array $options = []): ResponseInterface {
-            $this->assertSame('POST', $method);
-            $this->assertSame(self::URL, $url);
-            $this->assertSame(self::AUTH_HEADER, $options['normalized_headers']['authorization'][0]);
-            $this->assertSame('topic=https%3A%2F%2Fdemo.mercure.rocks%2Fdemo%2Fbooks%2F1.jsonld&data=Hi+from+Symfony%21&private=on&id=id&retry=3', $options['body']);
-
-            return new MockResponse('id');
-        });
-
-        // @phpstan-ignore-next-line
         $publisher = new Publisher(self::URL, new StaticJwtProvider(self::JWT), $httpClient);
         $id = $publisher(new Update(
             'https://demo.mercure.rocks/demo/books/1.jsonld',
@@ -89,43 +57,13 @@ class PublisherTest extends TestCase
         $this->assertSame('id', $id);
     }
 
-    public function testNetworkIssue()
-    {
-        $httpClient = new MockHttpClient(function (string $method, string $url, array $options = []): ResponseInterface {
-            $this->assertSame('POST', $method);
-            $this->assertSame(self::URL, $url);
-            $this->assertSame(self::AUTH_HEADER, $options['normalized_headers']['authorization'][0]);
-            $this->assertSame('topic=https%3A%2F%2Fdemo.mercure.rocks%2Fdemo%2Fbooks%2F1.jsonld&data=Hi+from+Symfony%21&private=on&id=id&retry=3', $options['body']);
-
-            throw new TransportException('Ops');
-        });
-
-        $provider = new StaticTokenProvider(self::JWT);
-        $hub = new Hub(self::URL, $provider);
-        $publisher = new Publisher($hub, $httpClient);
-
-        $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage('Failed to send an update.');
-
-        $publisher->publish(new Update(
-            'https://demo.mercure.rocks/demo/books/1.jsonld',
-            'Hi from Symfony!',
-            true,
-            'id',
-            null,
-            3
-        ));
-    }
-
     public function testInvalidJwt()
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('The provided JWT is not valid');
 
-        $provider = new StaticTokenProvider("invalid\r\njwt");
-        $hub = new Hub(self::URL, $provider);
-        $publisher = new Publisher($hub);
+        $publisher = new Publisher(self::URL, new StaticJwtProvider('foo'));
 
-        $publisher->publish(new Update('https://demo.mercure.rocks/demo/books/1.jsonld', 'Hi from Symfony!'));
+        $publisher(new Update('https://demo.mercure.rocks/demo/books/1.jsonld', 'Hi from Symfony!'));
     }
 }